requirements.txt
Flask
psycopg2
pillow
numpy
boto3
imutils
dlib
opencv-python
sklearn
scipy